requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
numpy>=1.26.0
httpx>=0.27.0
openai>=1.42.0